Upgrade 100ms spaces to working A/V room (video + screen share)
Makes the 100ms 'Go Live' option a real multiplayer video room and fixes
its core join bug.

- New shared HMSVideoRoom (video grid + screen share + controls), joined by
  per-room room_id_100ms or a room name.
- /spaces/hms/[id] now joins the room's own room_id_100ms with the caller's
  FID. Previously HMSRoom passed neither room id nor name (and a display name
  as userId), so every '100ms room' silently joined one shared default room
  and 403'd on the FID check. Created rooms are now isolated and joinable.
- AuditTestRoom refactored to a thin wrapper over HMSVideoRoom (one tested
  code path).

Everyone joins as 'speaker' so participants can talk + share video/screen;
no moderator role requested, so no token-route host gating. Stage-vs-room
listener semantics is a follow-up.
